requirements.txt
google-genai==1.47.0
pypdf==6.9.2
pytest==8.4.2
python-docx==1.2.0
spacy==3.8.11